reverted mission control to a working revision

+++ include/tasks/mission_control.h
@@ -7,7 +7,7 @@
 
 void mission_control(void) __attribute__((noreturn));
 
-int  update_turnout(int num, int state);
+int  update_turnout(short num, short  state);
 int  update_train_speed(int train, int speed);
 
 #endif
